Refactoring index.js
